// src/scrapers/extra/delta.d.ts
export interface DeltaResponse {
  badgesInfo: BadgesInfo
  tripOriginAirportCode: string
  tripOriginCityName: string
  tripDestinationAirportCode: string
  tripDestinationCityName: string
  currencyCode: string
  currencySymbol: string
  searchCriteria: SearchCriteria
  sessionTimeoutLimit: number
  sessionTimeoutModalLimit: number
  sessionTimeoutRemainingLimit: number
  customer: Customer
  customerReferenceId: string
  shopType: string
  tripType: string
  offerCall: string
  offerRequestId: string
  offerResponseId: string
  badgeInfoMap: BadgeInfoMap
  showCurrencyDisclaimer: boolean
  availableBrandId: AvailableBrandId[]
  page: Page
  itinerary: Itinerary[]
  filter: Filter
  sortOption: SortOption[]
  availableSearchOptionId: any[]
  links: Link3[]
  shopTypeOptions: ShopTypeOption2[]
  dynamicBannerRequest: DynamicBannerRequest
  nearbyAirports: NearbyAirports
  companionUpgrade: boolean
  airportMarketType: string
  metaData: MetaData
  amexCompanionCert: boolean
  hasTravelPolicy: boolean
}

export interface BadgesInfo {
  badges: Badge[]
}

export interface Badge {
  badgeText: BadgeText
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
  badgeURL?: string
  mobileBadgeIcon?: MobileBadgeIcon
  badgeIcon?: string
  travelStartDate?: string
  travelEndDate?: string
  searchStartDate?: string
  searchEndDate?: string
}

export interface BadgeText {
  hasMarkup: boolean
  text: string
}

export interface MobileBadgeIcon {
  type: string
  url: string
}

export interface SearchCriteria {
  request: Request
  fareBrandsByMiles: FareBrandsByMile[]
  fareBrandsByMoney: FareBrandsByMoney[]
  defaultsMap: DefaultsMap
  paxDefaults: PaxDefaults
  radiusDefaults: RadiusDefault[]
  shopTypeOptions: ShopTypeOption[]
  flexOptions: FlexOption[]
  journeyTypeOptions: JourneyTypeOption[]
  fareClassOptions: FareClassOption[]
  hasCoporateAgreement: boolean
  disableTripType: boolean
  disablePaxDropDown: boolean
  disableShowPriceIn: boolean
  disableBestFaresFor: boolean
  disableMeetingCode: boolean
  disableRefundableFares: boolean
  disableNearbyAirport: boolean
  disableEdocLink: boolean
  disableFlexibleDates: boolean
  useEdocDefaultPaxCount: boolean
  edocDefaultPaxCount: number
  edocMinPaxCount: number
  edocMaxPaxCount: number
  tripOriginCityName: string
  tripDestinationCityName: string
  corporateSearch: boolean
  meetingEventSearch: boolean
  edocSearch: boolean
}

export interface Request {
  bestFare: string
  action: string
  destinationAirportRadius: DestinationAirportRadius
  deltaOnlySearch: boolean
  meetingEventCode: string
  originAirportRadius: OriginAirportRadius
  passengers: Passenger[]
  searchType: string
  segments: Segment[]
  shopType: string
  tripType: string
  priceType: string
  priceSchedule: string
  awardTravel: boolean
  refundableFlightsOnly: boolean
  nonstopFlightsOnly: boolean
  datesFlexible: boolean
  flexCalendar: boolean
  flexAirport: boolean
  upgradeRequest: boolean
  corporateSMTravelType: string
  pageName: string
  cacheKey: string
  requestPageNum: string
  sortableOptionId: string
  selectedSolutions: any[]
  actionType: string
  initialSearchBy: InitialSearchBy
}
// ...
export interface FareBrandsByMile {
  brandID: string
  brandName: string
}

export interface FareBrandsByMoney {
  brandID: string
  brandName: string
}

export interface DefaultsMap {
  flexairportsearch: string
  flexCheck: string
  refundableFlightsOnly: string
  calendarsearch: string
  deltaAndPartners: string
  deltaOnlySearch: string
  nonstopFlightsOnly: string
}

export interface PaxDefaults {
  paxTypeList: PaxTypeList[]
}

export interface PaxTypeList {
  paxCode: string
  paxDesc: string
  defaultCount: number
  maximumPaxCount: number
}

export interface RadiusDefault {
  unit: string
  measure: number
  value: string
  isDefaulted: boolean
}

export interface ShopTypeOption {
  value: string
  isDefaulted: boolean
}

export interface FlexOption {
  value: string
  isDefaulted: boolean
}

export interface JourneyTypeOption {
  value: string
  isDefaulted: boolean
}

export interface FareClassOption {
  code: string
  description: string
}

export interface Customer {
  pwmDomesticFlag: boolean
  pwmInternationalFlag: boolean
  loggedInStatus: boolean
}

export interface BadgeInfoMap {
  LOWEST: Lowest
  NEWAIRPORT: Newairport
  IN_POLICY: InPolicy
  VTLFLT: Vtlflt
  Current_Flight: CurrentFlight
  NOT_ALLOWED: NotAllowed
  TRNSVC: Trnsvc
  FASTEST: Fastest
  OUT_POLICY: OutPolicy
  ORIGINALFLIGHT: Originalflight
  NEARBY: Nearby
  VIASAT: Viasat
  pendingGovtApproval: PendingGovtApproval
}

export interface Lowest {
  badgeText: BadgeText2
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
}

export interface BadgeText2 {
  hasMarkup: boolean
  text: string
}

export interface Newairport {
  badgeText: BadgeText3
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  badgeURL: string
  id: string
}

export interface BadgeText3 {
  hasMarkup: boolean
  text: string
}

export interface InPolicy {
  badgeText: BadgeText4
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
  mobileBadgeIcon: MobileBadgeIcon2
  badgeIcon: string
}

export interface BadgeText4 {
  hasMarkup: boolean
  text: string
}

export interface MobileBadgeIcon2 {
  type: string
  url: string
}

export interface Vtlflt {
  badgeText: BadgeText5
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
}

export interface BadgeText5 {
  hasMarkup: boolean
  text: string
}

export interface CurrentFlight {
  badgeText: BadgeText6
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  travelStartDate: string
  travelEndDate: string
  searchStartDate: string
  searchEndDate: string
  id: string
}

export interface BadgeText6 {
  hasMarkup: boolean
  text: string
}

export interface NotAllowed {
  badgeText: BadgeText7
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
  mobileBadgeIcon: MobileBadgeIcon3
  badgeIcon: string
}

export interface BadgeText7 {
  hasMarkup: boolean
  text: string
}

export interface MobileBadgeIcon3 {
  type: string
  url: string
}

export interface Trnsvc {
  badgeText: BadgeText8
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  badgeURL: string
  id: string
}

export interface BadgeText8 {
  hasMarkup: boolean
  text: string
}

export interface Fastest {
  badgeText: BadgeText9
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
}

export interface BadgeText9 {
  hasMarkup: boolean
  text: string
}

export interface OutPolicy {
  badgeText: BadgeText10
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
  mobileBadgeIcon: MobileBadgeIcon4
  badgeIcon: string
}

export interface BadgeText10 {
  hasMarkup: boolean
  text: string
}

export interface MobileBadgeIcon4 {
  type: string
  url: string
}

export interface Originalflight {
  badgeText: BadgeText11
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  travelStartDate: string
  travelEndDate: string
  searchStartDate: string
  searchEndDate: string
  id: string
}

export interface BadgeText11 {
  hasMarkup: boolean
  text: string
}

export interface Nearby {
  badgeText: BadgeText12
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
}

export interface BadgeText12 {
  hasMarkup: boolean
  text: string
}

export interface Viasat {
  badgeText: BadgeText13
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  travelStartDate: string
  travelEndDate: string
  searchStartDate: string
  searchEndDate: string
  id: string
}

export interface BadgeText13 {
  hasMarkup: boolean
  text: string
}

export interface PendingGovtApproval {
  badgeText: BadgeText14
  badgeIndicator: string
  badgeBgColor: string
  badgePriorityIndicator: string
  id: string
}

export interface BadgeText14 {
  hasMarkup: boolean
  text: string
}

export interface AvailableBrandId {
  brandId: string
  brandName: string
  refundable: boolean
  groupId: string
  deltaProduct: boolean
  productOfCarrier: string
  brandBckColor: string
  brandGradientColorStart: string
  brandGradientColorEnd: string
  genericSeatExperience: GenericSeatExperience
  isBasicEconomy: string
  priority: string
  brandGradientAngle: string
  displayBrandGradient: string
  showProductModal: string
  hideBrand: boolean
  defaultBrand: boolean
  brandGroupName?: string
  overlayIconUrl?: string
}

export interface GenericSeatExperience {
  bulletPromos: BulletPromos
  disclaimer?: Disclaimer
  header: Header
  description: Description
  icons?: Icon[]
}
// ...
export interface Page {
  resultCount: number
  num: number
  resultTotalCount: number
  highestResultCount: number
  lowestResultCount: number
}

export interface Itinerary {
  id: number
  fastest: boolean
  lowestFare: boolean
  seatReferenceId: string
  fare: Fare[]
  trip: Trip[]
  links: Link[]
  itineraryProduct: ItineraryProduct[]
  combinedBadgeSet: string[]
}

export interface Fare {
  miscFlightInfos: MiscFlightInfo[]
  fareId: number
  fareRulesCode: string[]
  ticketDesignator: string[]
  specialFareRulesCode: any[]
  segmentNbrFareRulesCode: any[]
  tax: any[]
  pointOfSaleCity: PointOfSaleCity
  baggageAllowance: BaggageAllowance
  fareRulesByFareBreak: any[]
  fareRestrictionByFlightSeg: any[]
  groupId: string
  currencySymbol: string
  seatSelectable: boolean
  showUpgradeEligible: boolean
  offered: boolean
  soldOut: boolean
  availableAtPostPurchase: boolean
  dominantSegmentBrandId: string
  solutionId: string
  brandByFlightLeg: BrandByFlightLeg[]
  fareRuleUrl: FareRuleUrl
  selectUrl: SelectUrl
  globalUpgradeCertificateStatus: boolean
  regionalUpgradeCertificateStatus: boolean
  offerId: string
  offerItemId: string
  upsellMetaData: UpsellMetaData
  fareBadges: any[]
  notAllowedByPolicy: boolean
  mixedBrandedProductExperience?: boolean
  cancelable?: boolean
  refundable?: boolean
  lowestFare?: boolean
  discountAvailable?: boolean
  webPriceAvailable?: boolean
  changeable?: boolean
  seatsAvailableCount?: number
  totalTaxAndFees?: TotalTaxAndFees
  basePrice?: BasePrice
  salesPrice?: SalesPrice
  totalPrice?: TotalPrice
  totalPriceByPTC?: TotalPriceByPtc
  fareCalcLine?: string
  paxInfo?: PaxInfo
  endorsement?: Endorsement[]
  totalPaxPrice?: TotalPaxPrice[]
  upgradeAvailable?: boolean
  multipleCos?: boolean
  multipleProducts?: boolean
  multipleCabins?: boolean
  payWithMilesEligible?: boolean
  upsell?: boolean
  selected?: boolean
  mixAndMatch?: boolean
  priority?: string
  unavailableForSale?: boolean
  fareProduct?: FareProduct[]
}
// ...
export interface Trip {
  id: number
  schedDepartLocalTs: string
  tripDepartureDateOutbound: string
  tripDepartureTime: string
  tripArrivalDate: string
  tripArrivalTime: string
  schedArrivalLocalTs: string
  originAirportCode: string
  originCity: string
  destAirportCode: string
  destCity: string
  stopCount: number
  nearByAirport: boolean
  differentDepartureAirport: boolean
  newAirport: boolean
  flightSegment: FlightSegment[]
  totalTripTime: TotalTripTime
  summarizedProducts: SummarizedProduct[]
  viewSeatUrls: ViewSeatUrl2[]
  newSubFleetType: boolean
  dominantFlightLeg: DominantFlightLeg
  originCnCd: string
  destinationCnCd: string
  tripProduct: TripProduct[]
}
// ...
export interface Link {
  rel: string
  href: string
  payload: Payload
}
// ...
export interface ItineraryProduct {
  product: any[]
}

export interface Filter {
  stopCount: number[]
  departAirportCode: any[]
  arrivalAirportCode: any[]
  layoverAirportCode: string[]
  layoverAirport: LayoverAirport[]
  flightDurationRange: FlightDurationRange
  connectionTimeRange: ConnectionTimeRange
  totalPriceRange: TotalPriceRange
  totalMilesRange: TotalMilesRange
  flightStop: FlightStop[]
}

export interface LayoverAirport {
  airportCode: string
  airportName: string
  cityName: string
  region: string
  nearByCities: any[]
}

export interface FlightDurationRange {
  min: number
  max: number
}

export interface ConnectionTimeRange {
  min: number
  max: number
}

export interface TotalPriceRange {
  min: number
  max: number
}

export interface TotalMilesRange {
  min: number
  max: number
}

export interface FlightStop {
  stopCount: number
  stop: string
  stopType: string
}

export interface SortOption {
  id: string
  desc: string
  link: Link2
}

export interface Link2 {
  rel: string
  href: string
  payload: Payload2
}
// ...
export interface Link3 {
  rel: string
  href: string
  payload: Payload3
}

export interface Payload3 {
  compareInputObjJson: string
}

export interface ShopTypeOption2 {
  rel: string
  href: string
  payload: Payload4
}
// ...
export interface DynamicBannerRequest {
  userdata: Userdata
  requestparam: Requestparam
}
// ...
export interface Requestparam {
  campaignId: string
}

export interface NearbyAirports {
  hasNearby: boolean
  nearByDestinationAirportCodes: any[]
  nearByOriginAirportCodes: any[]
}

export interface MetaData {
  offerRequestId: string
  offerResponseId: string
  solutionSetID: string
}